refactor wildcard handling
