+++ builtins/exit.c
@@ -6,7 +6,7 @@
 /*   By: lalex-ku <lalex-ku@42sp.org.br>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/05/12 14:02:34 by lalex-ku          #+#    #+#             */
-/*   Updated: 2022/06/26 20:06:26 by lalex-ku         ###   ########.fr       */
+/*   Updated: 2022/06/28 18:51:00 by lalex-ku         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
