+++ executes/execute_external.c
@@ -6,7 +6,7 @@
 /*   By: lalex-ku <lalex-ku@42sp.org.br>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/05/09 15:39:20 by lalex-ku          #+#    #+#             */
-/*   Updated: 2022/06/24 20:26:44 by lalex-ku         ###   ########.fr       */
+/*   Updated: 2022/06/28 18:51:00 by lalex-ku         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
