+++ executes/multiple_commands_utils.c
@@ -6,7 +6,7 @@
 /*   By: lalex-ku <lalex-ku@42sp.org.br>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/06/09 19:35:21 by lalex-ku          #+#    #+#             */
-/*   Updated: 2022/06/24 20:26:44 by lalex-ku         ###   ########.fr       */
+/*   Updated: 2022/06/28 18:51:00 by lalex-ku         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
