+++ utils/skip_quotes.c
@@ -1,19 +1,33 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   handle_expansions.c                                :+:      :+:    :+:   */
+/*   skip_quotes.c                                      :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: lalex-ku <lalex-ku@42sp.org.br>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2022/05/25 14:51:24 by lalex-ku          #+#    #+#             */
-/*   Updated: 2022/06/28 13:08:09 by lalex-ku         ###   ########.fr       */
+/*   Created: 2022/06/28 18:31:52 by lalex-ku          #+#    #+#             */
+/*   Updated: 2022/06/28 18:31:58 by lalex-ku         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "minishell.h"
 
-void	handle_expansions(char **input, t_env *minienv, int exit_status)
+int	skip_quotes(char *str)
 {
-	expand_exit_status(input, exit_status);
-	expand_variables(input, minienv);
+	int	i;
+
+	i = 0;
+	if (str[i] == '\'')
+	{
+		i++;
+		while (str[i] != '\'')
+			i++;
+	}
+	if (str[i] == '"')
+	{
+		i++;
+		while (str[i] != '"')
+			i++;
+	}
+	return (i);
 }
